pyproject.toml
[project]
name = "equipment_recommender_rag"
version = "0.1.0"
description = "Equipment recommendation using RAG or dense retrieval"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
package = false
